requirements.txt
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.28
langchain-experimental==0.3.3
langchain-openai==0.2.13
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.147
PyMuPDF==1.24.13
fastembed==0.4.2
python-dotenv==1.0.1
qdrant-client==1.12.1
pydantic==2.10.4
streamlit==1.41.1
PyAudio==0.2.14
deepgram-sdk==3.7.7
groq
audio-recorder-streamlit
langchain-groq